% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.r
\name{atl.clean}
\alias{atl.clean}
\title{Acoustic Tag Localization: Smooth detections}
\usage{
atl.clean(data.tags, sensitivity = 0.02, span = 0.5, plot = T)
}
\arguments{
\item{data.tags}{A object returned from \code{atl.label}}

\item{sensitivity}{Sensitivity of the reflection detection (radians).}

\item{span}{Parameter for LOESS smoothing.}

\item{plot}{Set to \code{T} to plot results with \code{ggplot}.}

\item{freq.range}{Vector holding minimum and maximum pulse periods to scan.}

\item{freq.resolution}{Vector holding the dt of the pulse period search.}
}
\value{

}
\description{
This will remove reflections and apply a loess smoothing to the dataset, interpolating missed points.
For detecting reflections, reflections are assumed to arrive shortly after the direct signal.
Any detection which has an arrival time within \code{sensitivity} an earlier detection is considered a reflection.
You can set \code{plot = T} to get a visaul overview of the reflection filtering.
}
