% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.r
\name{atl.label}
\alias{atl.label}
\title{Acoustic Tag Localization: Label tags}
\usage{
atl.label(data, start, duration, n.tags, maxDistance, sensitivity,
  freq.range = c(0.9, 1.1), freq.resolution = 1e-04, plot = F,
  pulsePeriodDistance.idx = 10, time.shift.sensitivity = 3)
}
\arguments{
\item{data}{data.frame returned from \code{read.HTI.RAT()}.}

\item{start}{Start time in seconds for analysis.}

\item{duration}{Duration in seconds for analysis.}

\item{n.tags}{Number of tags to analyse with unique pulse onset frequencies.}

\item{maxDistance}{The maximum distance the individual can be from the hydrophone (meters).
This is used to calculate the filter bandwidth.}

\item{sensitivity}{Bandwidth in radians around target frequency.
Increase this value, if you are missing detections due to the tag changing frequency over time.}

\item{freq.range}{A vector holding the range of possible tag frequencies}

\item{plot}{Show ggplot graphics for error checking segments.}
}
\description{
Acoustic Tag Localization: Label tags
}
