% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra.r
\name{circular.mean}
\alias{circular.mean}
\title{Caclulate the circular mean}
\usage{
circular.mean(rad, weight = c(), verbose = T)
}
\arguments{
\item{rad}{A vector of radian values.}

\item{weight}{A vector of weights for radian values}

\item{verbose}{Can be usefull for debugging code.}
}
\description{
Caclulate the circular mean
}
