% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{dataset_HTI}
\alias{dataset_HTI}
\title{Example dataset from HTI AR102 Acoustic Tag System}
\format{A data frame with 38120 rows and 10 variables:}
\usage{
dataset_HTI
}
\description{
Time of arrival information from an array of 4 hydrophones positioned for 3D localizations.
}
\keyword{datasets}
