% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localizing.r
\name{findc}
\alias{findc}
\title{Find speed of sound from hydrophone array}
\usage{
findc(c = 1500, sensorLocations, detections)
}
\arguments{
\item{c}{Your starting value for estimating the speed of sound.  Defaults to 1500m/s}

\item{sensorLocations}{A `n` x `3` array holding the xyz coordinates of the `n` hydrophones used in the setup.}

\item{detections}{A `source` x `hydrophone` array the detection times in seconds of the acoustic source.}
}
\value{
The estimated true speed of sound in m/s
}
\description{
By minimizing error within the sheprical interpolation time-of-arrival-localization results and assuming the provided hydrophone positions are correct, this function will estimate the true speed of sound in your setup.
This is usefull for validating the speed of sound before processing with YAPS.
}
