% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_handling.r
\name{HTI.meta}
\alias{HTI.meta}
\title{Scan HTI metadata in folder}
\usage{
HTI.meta(folder)
}
\arguments{
\item{folder}{}
}
\value{
data.frame of metadata
}
\description{
This function will scan all PAT files in a given folder and return the metadata accociated with them.
This can be used for filtering your files by recording time before analysing.
}
