% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{Pen_SeaBass}
\alias{Pen_SeaBass}
\title{Example Tracking data for testing localization methods}
\format{A set of lists (one for each dataset), where each list contains two dataframes: One for the positions over time, and another for the hydrophone locations.}
\usage{
Pen_SeaBass
}
\description{
Contains:
\begin{enumerate}
\item Pen_SeaBass: Seabass swimming in a small scale mesocosm setup.
\end{enumerate}
}
\keyword{datasets}
