% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.r
\name{plot.tagDetectionsPP}
\alias{plot.tagDetectionsPP}
\title{Plot filtered and smoothed acoustic tag detections}
\usage{
\method{plot}{tagDetectionsPP}(x, ...)
}
\arguments{
\item{x}{}
}
\description{
Plot filtered and smoothed acoustic tag detections
}
