% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_handling.r
\name{read.HTI.RAT}
\alias{read.HTI.RAT}
\title{Import raw data for HTI 291 Acoustic Tag Reviever}
\usage{
read.HTI.RAT(file, fs)
}
\arguments{
\item{file}{The path to .RAT file holding raw data from reciever.}

\item{fs}{The sample rate of the file.}
}
\value{
A dataframe containing all detected pulses from all hydrophones.
}
\description{
Import raw data for HTI 291 Acoustic Tag Reviever
}
\examples{
file <- 'HTI_Example_Data/AT16S012840914 (G8 4,0,48).RAT'
data.raw <- read.HTI.RAT(file)

}
