% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.r
\name{tagDetectionPulse}
\alias{tagDetectionPulse}
\title{Detection and Identification of Pulses with Variable Periods}
\usage{
tagDetectionPulse(data = data.raw, n.tags = 4, sensitivity = 0.2,
  window = 60 * 3, span = 0.3, plot = T, proof = F,
  pulsePeriodDistance.idx = 10, time.shift.sensitivity = 3)
}
\arguments{
\item{data}{Dataset returned from \code{read.HTI.RAT} function.}

\item{n.tags}{NUmber of tags with a unique pulse period.}

\item{sensitivity}{Sensitivity in radians for pulse period filtering.}

\item{window}{Window size (seconds)}

\item{span}{Amount of smoothing to apply to loess curve.}

\item{plot}{Show ggplot of smoothed data as the analysis runs.}

\item{proof}{Show verbose plots and pause analysis for each plot.
This is intended for helping choose initial analysis parameters.}

\item{time.shift.sensitivity}{A parameter used to detect shifts in pulse detections.
Higher values will make the analysis more sensitive to correcting time shifts.}

\item{pulsePeriodDistance}{The minimum distance (index units) between pulse period frequencies.}
}
\description{
For acoustic tag localization experiments where the unique tag identity is defined by the pulse period.
This is intended for processing datasets from experiments which used acoustically tagged fish surrounded by a hydrophone array.
}
\details{
This function will:
1. Automatically detect the periods of each unique tag.
2. Assign each pulse detection to the matching individual.
3. Filter out detected pulse reflections.
4. Apply loess smoothing to the resulting dataset.

To verify the parameters you chose, set \code{plot = T} to see a visual overview of the dataset as it's processed.
}
