% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.r
\name{TagFreq.clean}
\alias{TagFreq.clean}
\title{Clean outliers from TOA dataset}
\usage{
TagFreq.clean(detections, z = 1.96, plot = T)
}
\arguments{
\item{z}{The degree of filtering as defined by the guassian z-score.}

\item{plot}{Set to T to show results of cleaning.}

\item{detection.periods}{A data.frame returned from `TagFreq.label`.}
}
\description{
Clean outliers from TOA dataset
}
