% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.r
\name{TagFreq.detect}
\alias{TagFreq.detect}
\title{Detect pulse onset frequencies of acoustic tags}
\usage{
TagFreq.detect(detections, frequencies, fs, n.tags, plot = F,
  pulsePeriodDistance.idx = 10)
}
\arguments{
\item{detections}{A vector of detection times (seconds) returned from \code{read.HTI.RAT()}.}

\item{frequencies}{A vector of frequencies to test}

\item{fs}{Sample rate of \code{detections}}

\item{n.tags}{The number of tags with unique frequencies}

\item{plot}{Set to T to plot results with ggplot}
}
\description{
Detect pulse onset frequencies of acoustic tags
}
\details{
This function finds the unique tag frequencies in setups where each tagged individual has a unique pulse interval on their acoustic tag.
You can set \code{plot = T} for error checking.
The resulting graph will show the mean resultant length of the circular mean for each tested frequency.
The resulting plot should show a clear, sharp peak for each unique tag frequency.
}
