% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.r
\name{TagFreq.filter}
\alias{TagFreq.filter}
\title{Filter pulse onsets by frequency}
\usage{
TagFreq.filter(detections, frequencies, fs, plot = F,
  pulsePeriodDistance.idx = 10, sensitivity = 0.1,
  time.shift.sensitivity = 3)
}
\arguments{
\item{detections}{A vector of detection times (seconds) returned from \code{read.HTI.RAT()}.}

\item{frequencies}{A vector of pulse onset frequencies to filter.}

\item{fs}{Sample rate of \code{detections}}

\item{plot}{Set to T to plot results with ggplot}

\item{sensitivity}{Bandwidth in radians around target frequency.
Increase this value, if you are missing detections due to the tag changing frequency over time.}

\item{maxDistance}{The maximum distance the individual can be from the hydrophone (meters).
This is used to calculate the filter bandwidth.}
}
\value{

}
\description{
This functions returns the indicies of those pulse detections which match with the provided pulse period frequencies.
Set \code{plot = T} to show filtering results in ggplot.
}
