% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.r
\name{TagFreq.label}
\alias{TagFreq.label}
\title{Label detections according to pulse period}
\usage{
TagFreq.label(data, plot = F, sensitivity = NA)
}
\arguments{
\item{data}{data.frame with columns `c("Tag", "Hydrophone", "Seconds")`.}

\item{plot}{Show ggplot graphics for error checking segments.}

\item{sensitivity}{An optional value in radians for plotting the sensitivity bandwidth in the resulting ggplot.
This is just a approximate visual aid for selecting filter sensitivities for the `Freq.filter` function and should not be interpreted as exact.}
}
\description{
Label detections according to pulse period
}
