% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationTools.r
\name{toaSimError}
\alias{toaSimError}
\title{Simulate error on an array of time-of-arrival values}
\usage{
toaSimError(toa, reciever.fs = c(), hydrophoneMovement.sd = c(),
  p.detect = c())
}
\arguments{
\item{toa}{An `n`-by-`m` matrix or data.frame, holding the time-of-arrival values, in seconds.
The number of pings is `n`  while `m` is the number of hydrophones in the array.}

\item{reciever.fs}{The sample rate of the reciever (Hz) used to simulate toa error.}

\item{hydrophoneMovement.sd}{The standard deviation of the 0 centered Guassian distribution used to simulate toa error due to hydrophone movement.
A single value to be applied to all hydrophones, or a vector which equals the length of the number of hydrophones.}

\item{p.detect}{The probability (between 0 and 1) that a hydrophone will detect a pulse.
A single value to apply to all hydrophones, or a vector of values the same length as the number of hydrophones.}

\item{tag.positions}{An `n`-by-4 matrix or dataframe holding the x, y, z positions and time in seconds of the tag pulses where `n` is the number of pulses.}

\item{c}{Speed of sound through the medium (meters per second).}
}
\description{
The parameter `hydrophoneMovement.sd` will simulate spatial error by applying a guassian normal error value to the distance between the hydrophone and the tag.
This is intended to simulate random hydrophone movement in three dimentions which way be the result of wave action (in particular for hydrophones mounted on floating structures).
}
