% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationTools.r
\name{xyz2toa}
\alias{xyz2toa}
\title{Convert xyz positions to time of arrival values for a hydrophone array setup}
\usage{
xyz2toa(hydrophone.positions, tag.positions, c = 1500)
}
\arguments{
\item{hydrophone.positions}{An `n`-by-3 matrix or data.frame, holding the x, y, and z positions, in meters, of the hydrophones where `n`` is the number of hydrophones in the array.}

\item{tag.positions}{An `n`-by-4 matrix or dataframe holding the x, y, z positions and time in seconds of the tag pulses where `n` is the number of pulses.}

\item{c}{Speed of sound through the medium (meters per second).}
}
\description{
This can be used to either calculate time-of-arrival information from a given set of tag and hydrophone positions.
The parameters `reciever.fs`, `hydrophoneMovement.sd`, and `p.detect` can be used to simulated sources of error from the detection systems sample rate, hydrophone movement, and probabolity of detection, respectively.
}
\details{
The parameter `hydrophoneMovement.sd` will simulate spatial error by applying a guassian normal error value to the distance between the hydrophone and the tag.
This is intended to simulate random hydrophone movement in three dimentions which way be the result of wave action (in particular for hydrophones mounted on floating structures).
}
