% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.r
\name{TagFreq.detect}
\alias{TagFreq.detect}
\title{Detect pulse onset frequencies of acoustic tags}
\usage{
TagFreq.detect(detections, frequencies, n.tags, minGap, plot = F)
}
\arguments{
\item{detections}{A vector of detection times (seconds) from a single hydrophone.}

\item{frequencies}{A vector of frequencies to test.}

\item{n.tags}{The number of tags with unique frequencies.}

\item{plot}{Set to T to plot results with ggplot.}

\item{miNGap}{The minimum gap between unique tag frequencies (Hz).}
}
\description{
Detect pulse onset frequencies of acoustic tags
}
\details{
This function uses convolution to find the unique tag frequencies in setups where each tagged individual has a unique pulse interval on their acoustic tag.
You can set \code{plot = T} for error checking.
}
