% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.r
\name{TagFreq.filter}
\alias{TagFreq.filter}
\title{Filter pulse onsets by frequency}
\usage{
TagFreq.filter(detections, frequencies, sensitivity = 0.1, plot = F)
}
\arguments{
\item{detections}{A vector of detection times (seconds).}

\item{frequencies}{A vector of pulse onset frequencies (Hz) to filter.}

\item{sensitivity}{Bandwidth in radians around target frequency.}

\item{plot}{Set to T to plot results with ggplot
Increase this value, if you are missing detections due to the tag changing frequency over time.}
}
\description{
This functions returns the indicies of those pulse detections which match with the provided pulse period frequencies.
Set \code{plot = T} to show filtering results in ggplot.
}
